move warning into separate file
