+++ lib/dependencies/ImportParserPlugin.js
@@ -10,7 +10,7 @@ const ImportDependenciesBlock = require("./ImportDependenciesBlock");
 const ImportEagerDependency = require("./ImportEagerDependency");
 const ContextDependencyHelpers = require("./ContextDependencyHelpers");
 const UnsupportedFeatureWarning = require("../UnsupportedFeatureWarning");
-const WebpackError = require("../WebpackError");
+const CommentCompilationWarning = require("../CommentCompilationWarning");
 
 class ImportParserPlugin {
 	constructor(options) {
@@ -257,19 +257,6 @@ class ImportParserPlugin {
 				return true;
 			}
 		});
-	}
-}
-
-class CommentCompilationWarning extends WebpackError {
-	constructor(message, module, loc) {
-		super(message);
-
-		this.name = "CommentCompilationWarning";
-
-		this.module = module;
-		this.loc = loc;
-
-		Error.captureStackTrace(this, this.constructor);
 	}
 }
